unify Response data classes into ResponseDTO
