+++ app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/CharacterItemByIdRequestDTO.kt
@@ -1,0 +1,9 @@
+package com.jorgeSM.apimarvel.data.remote.models
+
+/*****
+ * Projecto: Api marvel
+ * From: com.jorgeSM.apimarvel.data.remote.models
+ * Create by Jorge S. Medina on 9/9/22 at 09:15
+ * More info: https://www.linkedin.com/in/jorge-s%C3%A1nchez-medina-bb7b7371/
+ *****/
+class CharacterItemByIdRequestDTO (val id: String, val hast: String)
