--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/Comics.kt
@@ -1,8 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class Comics(
-    val available: Long? = null,
-    val collectionURI: String? = null,
-    val items: List<ComicsItem>? = null,
-    val returned: Long? = null
-)
