--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/ComicsItem.kt
@@ -1,6 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class ComicsItem(
-    val resourceURI: String? = null,
-    val name: String? = null
-)
