--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/Data.kt
@@ -1,9 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class Data(
-    val offset: Long? = null,
-    val limit: Long? = null,
-    val total: Long? = null,
-    val count: Long? = null,
-    val results: List<Result>? = null
-)
