+++ app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/ListOfCharacterRequestDTO.kt
@@ -6,4 +6,4 @@ package com.jorgeSM.apimarvel.data.remote.models
  * Create by Jorge S. Medina on 9/9/22 at 09:13
  * More info: https://www.linkedin.com/in/jorge-s%C3%A1nchez-medina-bb7b7371/
  *****/
-class ListOfCharacterRequest(val hast: String)
+class ListOfCharacterRequestDTO(val hast: String)
