--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/Stories.kt
@@ -1,8 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class Stories(
-    val available: Long? = null,
-    val collectionURI: String? = null,
-    val items: List<StoriesItem>? = null,
-    val returned: Long? = null
-)
