--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/StoriesItem.kt
@@ -1,7 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class StoriesItem(
-    val resourceURI: String? = null,
-    val name: String? = null,
-    val type: String? = null
-)
