--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/Thumbnail.kt
@@ -1,6 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class Thumbnail(
-    val path: String? = null,
-    val extension: String? = null
-)
