--- app/src/main/java/com/jorgeSM/apimarvel/data/remote/models/URL.kt
@@ -1,6 +1,0 @@
-package com.jorgeSM.apimarvel.data.remote.models
-
-data class URL(
-    val type: String? = null,
-    val url: String? = null
-)
