--- app/src/main/java/com/jorgeSM/apimarvel/data/usecase/UseCase.kt
@@ -1,20 +1,0 @@
-package com.jorgeSM.apimarvel.data.usecase
-
-import kotlinx.coroutines.*
-
-/*****
- * Projecto: Api marvel
- * From: com.jorgeSM.apimarvel.data
- * Create by Jorge S. Medina on 9/9/22 at 09:36
- * More info: https://www.linkedin.com/in/jorge-s%C3%A1nchez-medina-bb7b7371/
- *****/
-interface UseCase<in T, out S> {
-//    fun invoke(
-//        scope: CoroutineScope,
-//        params: T? = null,
-//        onResult: (/*Either<"ErrorControl"*/, S>)-> Unit,
-//        dispatcherWorker: CoroutineDispatcher = Dispatchers.IO
-//    ){
-//        scope.launch { withContext(dispatcherWorker){/* onResult(run(params)*/} }
-//    }
-}
