+++ app/src/main/java/com/jorgeSM/apimarvel/domain/models/ListOfCharacterRequest.kt
@@ -1,9 +1,9 @@
-package com.jorgeSM.apimarvel.data.remote.models
+package com.jorgeSM.apimarvel.domain.models
 
 /*****
  * Projecto: Api marvel
  * From: com.jorgeSM.apimarvel.data.remote.models
- * Create by Jorge S. Medina on 9/9/22 at 09:15
+ * Create by Jorge S. Medina on 9/9/22 at 09:13
  * More info: https://www.linkedin.com/in/jorge-s%C3%A1nchez-medina-bb7b7371/
  *****/
-class CharacterItemByIdRequest ( val id: String, val hast: String)
+class ListOfCharacterRequest(val hast: String)
