+++ app/src/main/java/com/jorgeSM/apimarvel/presentation/viewmodel/DetailsViewModel.kt
@@ -1,10 +1,10 @@
-package com.jorgeSM.apimarvel.viewmodel
+package com.jorgeSM.apimarvel.presentation.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jorgeSM.apimarvel.data.remote.models.CharacterItemByIdRequest
+import com.jorgeSM.apimarvel.data.remote.models.CharacterItemByIdRequestDTO
 import com.jorgeSM.apimarvel.domain.usecase.GetCharacterByIdUC
 import com.jorgeSM.apimarvel.presentation.mapper.transformToVO
 import com.jorgeSM.apimarvel.presentation.modelVO.ResultVO
@@ -14,7 +14,7 @@ import kotlinx.coroutines.launch
 
 /*****
  * Projecto: Api marvel
- * From: com.jorgeSM.apimarvel.viewmodel
+ * From: com.jorgeSM.apimarvel.presentation.viewmodel
  * Create by Jorge S. Medina on 9/9/22 at 10:32
  * More info: https://www.linkedin.com/in/jorge-s%C3%A1nchez-medina-bb7b7371/
  *****/
@@ -32,7 +32,7 @@ class DetailsViewModel(
         requestJob?.cancel()
         requestJob = viewModelScope.launch(Dispatchers.IO) {
             getCharacterByIdUC(
-                CharacterItemByIdRequest(id, hash)
+                CharacterItemByIdRequestDTO(id, hash)
             )?.let {
                 it.data?.results?.map { result ->
                     _character.postValue(result.transformToVO())
